v 0.0.8 -added favorite to Quote -added TextLoading -added tooltip directive -added uppercase filter -added pagination -some other changes

+++ src/store/words.js
@@ -2,18 +2,29 @@ export default {
   state: {},
   mutations: {},
   actions: {
-    async addWord({ commit }, data) {
+    async addWord({ rootState, dispatch, commit }, data) {
       try {
-        const response = await _fetch("/words/add", {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify(data),
-        })
-        if (response.status === 409) return false
-        if (response.status === 200) return true
-        throw new Error()
+        const token = rootState.auth.accessToken
+        if (token) {
+          const response = await _fetch("/words/add", {
+            method: "POST",
+            headers: {
+              "Content-Type": "application/json",
+              Authorization: "bearer " + token,
+            },
+            body: JSON.stringify(data),
+          })
+          if (response.status === 200) return true
+          if (response.status === 401) {
+            const res = await dispatch("refreshTokens", {
+              action: "addWord",
+              actionData: data,
+            })
+            if (res) return true
+          }
+          if (response.status === 409) return false
+          throw new Error()
+        }
       } catch (e) {
         commit("setError", e)
         throw e
@@ -21,14 +32,14 @@ export default {
     },
     async searchWords({ commit }, data) {
       try {
-        const response = await _fetch("/words/search?word=" + data, {
+        const response = await _fetch("/words/search" + data, {
           method: "GET",
           headers: {
             "Content-Type": "application/json",
           },
         })
         if (response.status > 200) throw new Error()
-       const res = await response.json()
+        const res = await response.json()
         if (res.success === false) return true
         return res
       } catch (e) {
@@ -45,7 +56,7 @@ export default {
           },
         })
         if (response.status > 200) throw new Error()
-       const res = await response.json()
+        const res = await response.json()
         if (res.success === false) return true
         return res
       } catch (e) {
@@ -53,35 +64,58 @@ export default {
         throw e
       }
     },
-    async updateWord({ commit }, data) {
+    async updateWord({ rootState, dispatch, commit }, data) {
       try {
-        const response = await _fetch("/words/update", {
-          method: "PUT",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify(data),
-        })
-        if (response.status > 200) throw new Error()
-        if (response.status === 200) return await response.json()
-        throw new Error()
+        const token = rootState.auth.accessToken
+        if (token) {
+          const response = await _fetch("/words/update", {
+            method: "PUT",
+            headers: {
+              "Content-Type": "application/json",
+              Authorization: "bearer " + token,
+            },
+            body: JSON.stringify(data),
+          })
+          if (response.status === 200) return await response.json()
+
+          if (response.status === 401) {
+            const res = await dispatch("refreshTokens", {
+              action: "updateWord",
+              actionData: data,
+            })
+            if (res) return res
+          }
+          if (response.status > 200) throw new Error()
+          throw new Error()
+        }
       } catch (e) {
         commit("setError", e)
         throw e
       }
     },
-    async deleteWord({ commit }, data) {
+    async deleteWord({ rootState, dispatch, commit }, data) {
       try {
-        const response = await _fetch("/words/delete", {
-          method: "DELETE",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify(data),
-        })
-        if (response.status > 200) throw new Error()
-        if (response.status === 200) return true
-        throw new Error()
+        const token = rootState.auth.accessToken
+        if (token) {
+          const response = await _fetch("/words/delete", {
+            method: "DELETE",
+            headers: {
+              "Content-Type": "application/json",
+              Authorization: "bearer " + token,
+            },
+            body: JSON.stringify(data),
+          })
+          if (response.status === 200) return true
+          if (response.status === 401) {
+            const res = await dispatch("refreshTokens", {
+              action: "deleteWord",
+              actionData: data,
+            })
+            if (res) return res
+          }
+          if (response.status > 200) throw new Error()
+          throw new Error()
+        }
       } catch (e) {
         commit("setError", e)
         throw e
